requirements.txt
streamlit
st-cookies-manager
spotipy
openai
python-dotenv